+++ src/Controls/Controls/node_control.py
@@ -1,0 +1,44 @@
+import rclpy
+from rclpy.node import Node
+
+from std_msgs.msg import String
+
+
+class VehicleControllerNode(Node):
+    """
+    This class is a ROS2 node that controls the vehicle.
+    
+    """
+    def __init__(self):
+        super().__init__('Vehicle')
+        
+        self.get_logger().info("Vehicle Controller Node is Initializing...")
+        self.publisher_ = self.create_publisher(String, 'topic', 10)
+        timer_period = 0.5  # seconds
+        self.timer = self.create_timer(timer_period, self.timer_callback)
+        self.i = 0
+
+    def timer_callback(self):
+        msg = String()
+        msg.data = 'Hello World: %d' % self.i
+        self.publisher_.publish(msg)
+        self.get_logger().info('Publishing: "%s"' % msg.data)
+        self.i += 1
+
+
+def main(args=None):
+    rclpy.init(args=args)
+
+    vehicle_controller = VehicleControllerNode()
+
+    rclpy.spin(vehicle_controller)
+
+    # Destroy the node explicitly
+    # (optional - otherwise it will be done automatically
+    # when the garbage collector destroys the node object)
+    vehicle_controller.destroy_node()
+    rclpy.shutdown()
+
+
+if __name__ == '__main__':
+    main()
