+++ src/Controls/setup.py
@@ -20,8 +20,8 @@ setup(
     tests_require=['pytest'],
     entry_points={
         'console_scripts': [
-            'talker = py_pubsub.publisher_member_function:main',
-            'listener = py_pubsub.subscriber_member_function:main',
+            'controller = controls.node_control:main',
+            'listener = controls.subscriber_member_function:main',
         ],
     },
 )
