+++ src/Navigation/package.xml
@@ -12,6 +12,8 @@
   <test_depend>ament_pep257</test_depend>
   <test_depend>python3-pytest</test_depend>
 
+  <exec_depend>custom_interfaces</exec_depend>
+
   <export>
     <build_type>ament_python</build_type>
   </export>
