+++ src/custom_interfaces/msg/HelperLatLon.msg
@@ -1,0 +1,2 @@
+float32 latitude # Unit: degrees
+float32 longitude # Unit: degrees
