+++ src/custom_interfaces/msg/Location.msg
@@ -1,0 +1,1 @@
+HelperLatLon[] lat_lon
