+++ src/custom_interfaces/msg/MotorPower.msg
@@ -1,0 +1,1 @@
+float32 power # power of the signal sent to the motor
