+++ src/custom_interfaces/msg/NewHeading.msg
@@ -1,0 +1,1 @@
+float32 new_heading # units in degrees
