+++ src/custom_interfaces/msg/SonarSensor.msg
@@ -1,0 +1,1 @@
+float32 sonar_data # Sonar data in arbitary units
