+++ src/custom_interfaces/msg/Steering.msg
@@ -1,0 +1,1 @@
+float32 steering # steering angle in degrees
